fix(ast): fix name of `AstBuilder` method for `Expression::V8IntrinsicExpression` (#23766)
`ast_tools` codegen handles of converting types names to snake case. It ensures that `V8IntrinsicExpression` is converted to `v8_intrinsic_expression` not `v_8_intrinsic_expression`.

However, this special case was missed out when converting enum variant names.

Fix this, which results in `AstBuilder::expression_v_8_intrinsic` being corrected to `expression_v8_intrinsic`.
